Conversion to double fix.

+++ cunumeric/module.py
@@ -6437,7 +6437,7 @@ def histogram(
                                               for k in range(0, num_elems)]),
                              dtype=float)
     else:
-        bins_array = asarray(bins)
+        bins_array = asarray(bins).astype(float)
         num_intervals = bins_array.shape[0] - 1
 
     if x.ndim != 1:
@@ -6450,7 +6450,7 @@ def histogram(
         # bc/ of hist ndarray inputs(), below;
         # needs to be handled here:
         #
-        weights = ones(x.shape, dtype=x.dtype)
+        weights = ones(x.shape, dtype=float)
 
     hist = ndarray(
         (num_intervals,),
